requirements.txt
PyQt5==5.15.9
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
matplotlib==3.7.2
yfinance==0.2.28
msal==1.24.0
openpyxl==3.1.2
seaborn==0.12.2